requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy==2.1.3
